Added checktime
Also added a dump file, though that was mainly in use on Windows machine. As soon as RFID works, this whole file is changing to an update query because since server can't ping client, server will log a new med under current user and wait until medicines are viewed, and then will prompt user on android side "hey this is a new medicine wanna fill in some details?".