Refactored tables
Added a few fields.
User: (hashed) password and salt. Password to be necessary only on access from client.

Medicine: Purpose, reminded, stolen, new, inout. lastout. Reminded is used so if user needs to be reminded, they will be reminded again the next day if they forget again. Stolen is if wrong/no fingerprint, but take with grain of salt. New is true on a new medicine and prompts user to enter medicine info. Inout determines whether medicine is considered "in" or "out". Lastout is time that med was taken out. If med is out too long, user will be reminded to put their medicine away.

CheckTime: Added date for the longer time periods such as weeks or months.